+++ tests/testthat/test-collapsing.R
@@ -26,7 +26,7 @@ test_that("collapsing with formatting", {
 
 test_that("collapsing with formatting, custom seps", {
   expect_snapshot(local({
-    cli_div(theme = list(div = list(vec_sep = " ... ")))
+    cli_div(theme = list(div = list("vec-sep" = " ... ")))
     pkgs <- paste0("pkg", 1:5)
     cli_text("Packages: {.pkg {pkgs}}.")
   }))
@@ -36,7 +36,7 @@ test_that("collapsing a cli_vec", {
   expect_snapshot({
     pkgs <- cli_vec(
       paste0("pkg", 1:5),
-      style = list(vec_sep = " & ", vec_last = " & ")
+      style = list("vec-sep" = " & ", "vec-last" = " & ")
     )
     cli_text("Packages: {pkgs}.")
   })
@@ -44,10 +44,10 @@ test_that("collapsing a cli_vec", {
 
 test_that_cli(configs = c("plain", "ansi"), "collapsing a cli_vec with styling", {
   expect_snapshot(local({
-    cli_div(theme = list(body = list(vec_sep = " ... ")))
+    cli_div(theme = list(body = list("vec-sep" = " ... ")))
     pkgs <- cli_vec(
       paste0("pkg", 1:5),
-      style = list(vec_sep = " & ", vec_last = " & ", color = "blue")
+      style = list("vec-sep" = " & ", "vec-last" = " & ", color = "blue")
     )
     cli_text("Packages: {pkgs}.")
   }))